# controlplane/gen/.openapi-3.0.yaml
openapi: 3.0.3
info:
  title: LangSmith Deployment
  version: 0.1.0
  description: "LangGraph Cloud-compatible API surface implemented by duragraph.\n\
    \nNote: paths are documented as bare (`/assistants`, `/threads`, `/runs`,\netc.)\
    \ to mirror the upstream LangGraph spec. The duragraph engine\nserves these paths\
    \ under the `/api/v1` base \u2014 declared via the\n`servers:` block below. Clients\
    \ should construct request URLs as\n`<server.url><path>`, e.g. `http://localhost:8081/api/v1/runs`.\n"
servers:
- url: http://localhost:8081/api/v1
  description: Local development server
- url: https://api.duragraph.ai/api/v1
  description: "Production server (placeholder \u2014 replace with deployment host)"
...
paths:
  /assistants:
    post:
      tags:
      - Assistants
      summary: Create Assistant
      description: 'Create an assistant.


        An initial version of the assistant will be created and the assistant is set
        to that version. To change versions, use the `POST /assistants/{assistant_id}/latest`
        endpoint.'
      operationId: create_assistant_assistants_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssistantCreate'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assistant'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /assistants/search:
    post:
      tags:
      - Assistants
      summary: Search Assistants
      description: 'Search for assistants.


        This endpoint also functions as the endpoint to list all assistants.'
      operationId: search_assistants_assistants_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssistantSearchRequest'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Assistant'
                type: array
                title: Response Search Assistants Assistants Search Post
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /assistants/count:
    post:
      tags:
      - Assistants
      summary: Count Assistants
      description: Get the count of assistants matching the specified criteria.
      operationId: count_assistants_assistants_count_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssistantCountRequest'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: integer
                title: Count
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /assistants/{assistant_id}:
    get:
      tags:
      - Assistants
      summary: Get Assistant
      description: Get an assistant by ID.
      operationId: get_assistant_assistants__assistant_id__get
      parameters:
      - description: The ID of the assistant.
        required: true
        schema:
          type: string
          format: uuid
          title: Assistant ID
          description: The ID of the assistant.
        name: assistant_id
        in: path
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assistant'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - Assistants
      summary: Delete Assistant
      description: 'Delete an assistant by ID.


        All versions of the assistant will be deleted as well.'
      operationId: delete_assistant_assistants__assistant_id__delete
      parameters:
      - description: The ID of the assistant.
        required: true
        schema:
          type: string
          format: uuid
          title: Assistant ID
          description: The ID of the assistant.
        name: assistant_id
        in: path
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema: {}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - Assistants
      summary: Patch Assistant
      description: Update an assistant.
      operationId: patch_assistant_assistants__assistant_id__patch
      parameters:
      - description: The ID of the assistant.
        required: true
        schema:
          type: string
          format: uuid
          title: Assistant ID
          description: The ID of the assistant.
        name: assistant_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssistantPatch'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assistant'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /assistants/{assistant_id}/graph:
    get:
      tags:
      - Assistants
      summary: Get Assistant Graph
      description: Get an assistant by ID.
      operationId: get_assistant_graph_assistants__assistant_id__graph_get
      parameters:
      - description: The ID of the assistant.
        required: true
        schema: {}
        name: assistant_id
        in: path
      - description: Include graph representation of subgraphs. If an integer value
          is provided, only subgraphs with a depth less than or equal to the value
          will be included.
        required: false
        schema:
          title: Xray
          default: false
          description: Include graph representation of subgraphs. If an integer value
            is provided, only subgraphs with a depth less than or equal to the value
            will be included.
        name: xray
        in: query
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    type: object
                  type: array
                type: object
                title: Response Get Assistant Graph Assistants  Assistant Id  Graph
                  Get
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /assistants/{assistant_id}/versions:
    post:
      tags:
      - Assistants
      summary: Get Assistant Versions
      description: Get all versions of an assistant.
      operationId: get_assistant_versions_assistants__assistant_id__versions_get
      parameters:
      - description: The ID of the assistant.
        required: true
        schema:
          type: string
          format: uuid
          title: Assistant Id
          description: The ID of the assistant.
        name: assistant_id
        in: path
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Assistant'
                type: array
                title: Response Search Assistants Assistants Search Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /assistants/{assistant_id}/latest:
    post:
      tags:
      - Assistants
      summary: Set Latest Assistant Version
      description: Set the latest version for an assistant.
      operationId: set_latest_assistant_version_assistants__assistant_id__versions_post
      parameters:
      - description: The ID of the assistant.
        required: true
        schema:
          type: string
          format: uuid
          title: Assistant Id
          description: The ID of the assistant.
        name: assistant_id
        in: path
      - description: The version to change to.
        required: true
        schema:
          type: integer
          title: Version
          description: The version of the assistant to change to.
        name: version
        in: query
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assistant'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Assistant:
      properties:
        assistant_id:
          type: string
          format: uuid
          title: Assistant Id
          description: The ID of the assistant.
        graph_id:
          type: string
          title: Graph Id
          description: The ID of the graph.
        config:
          properties:
            tags:
              items:
                type: string
              type: array
              title: Tags
            recursion_limit:
              type: integer
              title: Recursion Limit
            configurable:
              type: object
              title: Configurable
          type: object
          title: Config
          description: The assistant config.
        context:
          type: object
          title: Context
          description: Static context added to the assistant.
        created_at:
          type: string
          format: date-time
          title: Created At
          description: The time the assistant was created.
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: The last time the assistant was updated.
        metadata:
          type: object
          title: Metadata
          description: The assistant metadata.
        version:
          type: integer
          title: Version
          description: The version of the assistant
        name:
          type: string
          title: Assistant Name
          description: The name of the assistant
        description:
          type: string
          title: Assistant Description
          description: The description of the assistant
          nullable: true
      type: object
      required:
      - assistant_id
      - graph_id
      - config
      - created_at
      - updated_at
      - metadata
      title: Assistant
    AssistantCreate:
      properties:
        assistant_id:
          type: string
          format: uuid
          title: Assistant Id
          description: The ID of the assistant. If not provided, a random UUID will
            be generated.
        graph_id:
          type: string
          title: Graph Id
          description: The ID of the graph the assistant should use. The graph ID
            is normally set in your langgraph.json configuration.
        config:
          type: object
          title: Config
          description: Configuration to use for the graph. Useful when graph is configurable
            and you want to create different assistants based on different configurations.
        context:
          type: object
          title: Context
          description: Static context added to the assistant.
        metadata:
          type: object
          title: Metadata
          description: Metadata to add to assistant.
        if_exists:
          type: string
          enum:
          - raise
          - do_nothing
          title: If Exists
          description: How to handle duplicate creation. Must be either 'raise' (raise
            error if duplicate), or 'do_nothing' (return existing assistant).
          default: raise
        name:
          type: string
          title: Name
          description: The name of the assistant. Defaults to 'Untitled'.
        description:
          type: string
          title: Description
          description: The description of the assistant. Defaults to null.
          nullable: true
      type: object
      required:
      - graph_id
      title: AssistantCreate
      description: Payload for creating an assistant.
    AssistantPatch:
      properties:
        graph_id:
          type: string
          title: Graph Id
          description: The ID of the graph the assistant should use. The graph ID
            is normally set in your langgraph.json configuration. If not provided,
            assistant will keep pointing to same graph.
        config:
          type: object
          title: Config
          description: Configuration to use for the graph. Useful when graph is configurable
            and you want to update the assistant's configuration.
        context:
          type: object
          title: Context
          description: Static context added to the assistant.
        metadata:
          type: object
          title: Metadata
          description: Metadata to merge with existing assistant metadata.
        name:
          type: string
          title: Name
          description: The new name for the assistant. If not provided, assistant
            will keep its current name.
        description:
          type: string
          title: Description
          description: The new description for the assistant. If not provided, assistant
            will keep its current description.
      type: object
      title: AssistantPatch
      description: Payload for updating an assistant.
...
    AssistantSearchRequest:
      properties:
        metadata:
          type: object
          title: Metadata
          description: Metadata to filter by. Exact match filter for each KV pair.
        graph_id:
          type: string
          title: Graph Id
          description: The ID of the graph to filter by. The graph ID is normally
            set in your langgraph.json configuration.
        limit:
          type: integer
          title: Limit
          description: The maximum number of results to return.
          default: 10
          minimum: 1
          maximum: 1000
        offset:
          type: integer
          title: Offset
          description: The number of results to skip.
          default: 0
          minimum: 0
        sort_by:
          type: string
          enum:
          - assistant_id
          - created_at
          - updated_at
          - name
          - graph_id
          title: Sort By
          description: The field to sort by.
        sort_order:
          type: string
          enum:
          - asc
          - desc
          title: Sort Order
          description: The order to sort by.
        select:
          type: array
          items:
            type: string
            enum:
            - assistant_id
            - graph_id
            - name
            - description
            - config
            - context
            - created_at
            - updated_at
            - metadata
            - version
          title: Select
          description: Specify which fields to return. If not provided, all fields
            are returned.
      type: object
      title: AssistantSearchRequest
      description: Payload for listing assistants.
    AssistantCountRequest:
      properties:
        metadata:
          type: object
          title: Metadata
          description: Metadata to filter by. Exact match filter for each KV pair.
        graph_id:
          type: string
          title: Graph Id
          description: The ID of the graph to filter by. The graph ID is normally
            set in your langgraph.json configuration.
      type: object
      title: AssistantCountRequest
      description: Payload for counting assistants.
...
    ErrorResponse:
      type: string
      title: ErrorResponse
      description: Error message returned from the server
...